requirements.txt
django==5.1.3
djangorestframework==3.15.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10
django-filter==24.3
drf-yasg==1.21.8
coverage==7.6.9
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0